requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
bidict==0.23.1
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
Django==4.2.27
django-ckeditor==6.7.3
django-htmx==1.27.0
django-js-asset==3.1.2
docutils==0.22.3
fastapi==0.124.2
frozenlist==1.8.0
h11==0.16.0
htmx==0.0.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown2==2.5.4
MarkupSafe==3.0.3
multidict==6.7.0
nicegui==3.4.0
orjson==3.11.5
pillow==12.0.0
propcache==0.4.1
pydantic==1.10.24
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.15.0
PyYAML==6.0.3
simple-websocket==1.1.0
sqlparse==0.5.4
starlette==0.50.0
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0
